Refactor redundant logic into a utils file
