+++ lametro/templates/lametro/legislation.html
@@ -10,8 +10,8 @@
             <div class="col-sm-12">
                 <h1 class="bill-page-title">
                     <br class="non-mobile-only" />
-                    {% if legislation.ocr_full_text|format_full_text|parse_subject %}
-                        {{ legislation.identifier}} - {{ legislation.ocr_full_text| format_full_text | parse_subject}}
+                    {% if legislation.ocr_full_text|prepare_title %}
+                        {{ legislation.identifier}} - {{ legislation.ocr_full_text|prepare_title }}
                     {% else %}
                         {{ legislation.identifier}}
                     {% endif %}
@@ -73,7 +73,7 @@
 
                 <div class='panel panel-default'>
                     <div class='panel-body'>
-                        {{ legislation.ocr_full_text| format_full_text |safe }}
+                        {{ legislation.ocr_full_text| prepare_title |safe }}
                     </div>
                 </div>
